Refactored API like things from experiemt and settings
